+++ src/main/java/zhang/lao/console/service/ConsoleSysRoleService.java
@@ -6,9 +6,9 @@ import zhang.lao.console.model.bootstrapQ.QTree;
  * Created by tech6 on 2016/8/2.
  */
 public interface ConsoleSysRoleService {
-    String getSelectRoleHtmlByUserId(Long id);
-    QTree getRoleNavJson(Long role_id,Long firstNavId);
-    void updateRoleNavByNavIdAndRoleId(String[] navs, Long role_id);
-    void updateUserRole(String[] role_ids, Long user_id);
-    boolean containNav(Long nav_id, Long role_id);
+    String getSelectRoleHtmlByUserId(Integer id);
+    QTree getRoleNavJson(Integer role_id,Integer firstNavId);
+    void updateRoleNavByNavIdAndRoleId(String[] navs, Integer role_id);
+    void updateUserRole(String[] role_ids, Integer user_id);
+    boolean containNav(Integer nav_id, Integer role_id);
 }
